requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
Pillow>=9.5.0
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0
jupyter>=1.0.0
